==> scripts/code/03_CSI_online_aphasia_spoken_descriptives_final.Rmd <==
---
title: '03 CSI online aphasia: Spoken - Descriptives'
author: "Kirsten Stark"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Load packages

```{r load_packages}
library(dplyr)
library(tidyr)

rm(list = ls())
```


## Load and preprocess data

```{r load_data}
options( "encoding" = "UTF-8" )

# input
input <- "aphasia_final.csv"

# load data
df <- read.csv(here::here("data", "transient_data_files", input), 
               sep = ",",  na = "")
```

## Duration of the experiment

```{r duration}
print("Total duration, not outlier corrected")
# mean(df$timetotal, na.rm=TRUE) # 33.13 min
mean(df$time_correct)
sd(df$timetotal) # 21.04 min
range(df$timetotal) # 15 to 126 min

print("Total duration, split by session")

df %>% group_by(type,session) %>% summarise(mean = mean(timetotal),
                                            median=median(timetotal),
                                       sd = sd(timetotal), 
                                       min = min(timetotal), 
                                       max = max(timetotal))
```


## Description of participants

Gender: 

```{r gender}
(df %>% mutate(gender_char = case_when(gender == 1 ~ "female", 
                                     gender == 2 ~ "male")) %>% 
  group_by(type, gender_char) %>% count(gender_char) %>% 
   mutate(n=n/160/3) -> gender)
 # 1 = female, 2 = male, 3 = diverse

print("percentage female:")
sum(df$gender == 1)/nrow(df)
```

Age: 

```{r age}
print('age:')
df %>% group_by(type) %>% 
  summarise(mean=round(mean(age),2), sd=round(sd(age),2), min=min(age), max=max(age))
```

Handedness: 

```{r handedness}
# 1 = left handed, 2 = right handed, 3 = ambidexter/both
(handedness <- df %>% mutate(handedness_char = case_when(handedness == 1 ~ "left-handed", 
                                     handedness == 2 ~ "right-handed", 
                                     handedness==3 ~ "ambidexter/both")) %>% 
  group_by(type) %>% count(handedness_char))

print("percentage right-handed:")
print('PWA')
sum(df$handedness[df$type=="PWA"] == 2)/nrow(df[df$type=="PWA",])
print('control')
sum(df$handedness[df$type=="control"] == 2)/nrow(df[df$type=="control",])
```


Mother tongue (experiment was restricted to native German speakers): This seems to have worked

```{r}
table(df$language) # 1 = yes (mother tongue is German), 2 = no
```



## Attention checks
*1) Item vs. non-item*

```{r item vs non-item}
## Item vs. non-item
# CH01_01 (Taube), CH01_02 (Apfel), CH02_01 (Luftballon) and CH02_02 (Biene) are items and 2 should be selected,
# CH01_03 (Radio), CH01_04 (Sparschwein), CH02_03 (Laptop) and CH02_04 (Wattestäbchen) are non-items and 1 should not be selected
## Did participants cheat
# CH03 = 1 - yes, I worked through it till the end,
# CH03 = 2 - no, I stopped or cheated midway
# CH03 = -9 - no answer

attcheck <- data.frame(subject = unique(df$subject)) 

df <- df %>% mutate(itemvsnonitem1 = 
    case_when(CH01_01==2 & CH01_02==2 & CH01_03==1 & CH01_04==1 ~2, # all correct
              CH01_01==2 | CH01_02==2 ~1, # one correct
              CH01_01!=2 & CH01_02!=2 ~0)) %>% # none correct
  dplyr::mutate(itemvsnonitem2 = 
    case_when(CH02_01==2 & CH02_02==2 & CH02_03==1 & CH02_04==1 ~2,
              CH02_01==2 | CH02_02==2 ~1, 
              CH02_01!=2 & CH02_02!=2 ~0))
df %>% group_by(type, session) %>% count(itemvsnonitem1) %>% mutate(n=n/160)
df %>% group_by(type, session) %>% count(itemvsnonitem2) %>% mutate(n=n/160)
# table(df$itemvsnonitem1)/160
# table(df$itemvsnonitem2)/160

```

All had at leas one item selected correctly in the attention test

*2) Cheating*

```{r cheating}
df <- df %>% mutate(CH03 = case_when(CH03 == 1 ~ 
                                       " Ja, ich habe alles bis zum Ende bearbeitet.",
                                             CH03 == 2 ~ 
                                       "Nein, ich habe zwischendurch aufgehoert oder geschummelt."))
df %>% group_by(type, session) %>% count(CH03) %>% mutate(n=n/160)
table(df$CH03)/160
```

## Comments
Comments don't indicate any problems that should lead to participant exclusion: 

```{r comments}
table(df$comments)/160
```

## Arrays

```{r}
table(df$array)/160
x <- df %>% group_by(subject, session) %>% count(array) %>% mutate(n=n/160)
array_rep = 0
for(i in 2:nrow(x)){
  if(x$subject[i-1] == x$subject[i] & x$array[i-1] == x$array[i]){
    array_rep = array_rep+1
  }
}
print(paste0(array_rep, " time an array was repeated within the same participant")) 

df %>% group_by(type, session) %>% count(array) %>% mutate(n=n/160) %>% arrange(array)
df %>% count(array) %>% mutate(n=n/160)
```

## Get an overview of the computer setups
Information was read-out automatically from SoSci Survey
SD22_PRV (provider):
1 = Android
2 = Apple
3 = BlackBerry
4 = PlayBook
5 = Kindle
6 = Microsoft
-1 = Sonstige oder unbekannt

SD22_OS (OS):
10 = Windows NT
11 = Windows Vista
12 = Windows 7
13 = Windows 8
15 = Windows 10
20 = Linux
30 = MaxOS X
40 = iOS
50 = Android
-1 = Sonstiges
-2 = unbekannt

SD22_BID (Browser-Identification (HTTP_USER_AGENT))

SD22_BNM (Browser)
1 = Android WebView
2 = Chrome
3 = Chromium
4 = Coast
5 = Firefox
6 = IE
7 = Opera
8 = Safari
9 = Edge
10 = Safari WebView
11 = IE (mobile)
12 = Edge (mobile)
-1 = Sonstiges
-2 = Default Browser
-3 = unbekannt

SD22_BVS (Browser-version)

SD22_FmF (Format)
1 = Computer
2 = Fernsehgerät
3 = Tablet
4 = Mobilgerät
5 = Smartphone
-2 = unbekannt

SD22_ScW (screen width [Pixel])
SD22_ScH (screen [Pixel]

SD22_QnW (questionnaire width [Pixel])


```{r}
#colnames(df)

df <- df %>% mutate(SD22_OS=as.character(as.numeric(SD22_OS))) %>%
  mutate(os_system = case_when(
  SD22_OS==10  ~ "Windows NT",
  SD22_OS==11  ~ "Windows Vista",
  SD22_OS==12  ~ "Windows 7",
  SD22_OS==13  ~ "Windows 8",
  SD22_OS==15  ~ "Windows 10",
  SD22_OS==20  ~ "Linux",
  SD22_OS==30  ~ "MaxOS X",
  SD22_OS==40  ~ "iOS",
  SD22_OS==50  ~ "Android",
  SD22_OS==-1  ~ "other",
  SD22_OS==-2  ~ "unknown")) %>%
  mutate(os_system_raw=case_when(
  SD22_OS==10  ~ "Windows",
  SD22_OS==11  ~ "Windows",
  SD22_OS==12  ~ "Windows",
  SD22_OS==13  ~ "Windows",
  SD22_OS==15  ~ "Windows",
  SD22_OS==20  ~ "Linux",
  SD22_OS==30  ~ "Apple",
  SD22_OS==40  ~ "Apple",
  SD22_OS==50  ~ "Android",
  SD22_OS==-1  ~ "other",
  SD22_OS==-2  ~ "unknown"))
df %>% group_by(os_system) %>% count() %>% mutate(n=n/40/3) 
df %>% group_by(os_system_raw) %>% count() %>% mutate(n=n/40/3) 

# df <- df %>% mutate(SD22_PRV=as.character(as.numeric(SD22_PRV))) %>%
#   mutate(provider = case_when(
#   SD22_PRV==1  ~ "Android",
#   SD22_PRV==2  ~ "Apple",
#   SSD22_PRV==3  ~ "BlackBerry",
#   SD22_PRV==4  ~ "PlayBook",
#   SD22_PRV==5  ~ "Kindle",
#   SD22_PRV==6  ~ "Microsoft",
#   SD22_PRV==-1  ~ "other or unknown")) 
# df %>% group_by(provider) %>% count() %>% mutate(n=n/40/3) 


#df %>% group_by(SD22_BID) %>% count() %>% mutate(n=n/40/3) 

df <- df %>% mutate(SD22_BNM=as.character(as.numeric(SD22_BNM))) %>%
  mutate(browser = case_when(
  SD22_BNM==1  ~ "Android WebView",
  SD22_BNM==2  ~ "Chrome",
  SD22_BNM==3  ~ "Chromium",
  SD22_BNM==4  ~ "Coast",
  SD22_BNM==5  ~ "Firefox",
  SD22_BNM==6  ~ "IE",
  SD22_BNM==7  ~ "Opera",
  SD22_BNM==8  ~ "Safari",
  SD22_BNM==9  ~ "Edge",
  SD22_BNM==10  ~ "Safari WebView",
  SD22_BNM==11  ~ "IE (mobile)",
  SD22_BNM==12  ~ "Edge (mobile)",
  SD22_BNM==-1  ~ "other",
  SD22_BNM==-2  ~ "default browser",
  SD22_BNM==-3  ~ "unknown"))

#df %>% group_by(SD22_BVS) %>% count() %>% mutate(n=n/40/3) # browser version

df <- df %>% mutate(SD22_FmF=as.character(as.numeric(SD22_FmF))) %>%
  mutate(system_format = case_when(
  SD22_FmF==1  ~ "Computer",
  SD22_FmF==2  ~ "Television",
  SD22_FmF==3  ~ "Tablet",
  SD22_FmF==4  ~ "Mobile device",
  SD22_FmF==5  ~ "Smartphone",
  SD22_BNM==-2  ~ "unknown"))
df %>% group_by(system_format) %>% count() %>% mutate(n=n/40/3) 

df <- df %>% mutate(screen_width=as.character(as.numeric(SD22_ScW)))
df %>% group_by(screen_width) %>% count() %>% mutate(n=n/40/3) 
df <- df %>% mutate(screen_height=as.character(as.numeric(SD22_ScH)))
df %>% group_by(screen_height) %>% count() %>% mutate(n=n/40/3) 
df <- df %>% mutate(questionnaire_width=as.character(as.numeric(SD22_QnW)))
df %>% group_by(questionnaire_width) %>% count() %>% mutate(n=n/40/3) 



(df %>% group_by(subject, session) %>% select(SD22_BID, os_system_raw,
                                             os_system, browser,
                                             system_format,
                                             screen_width, screen_height,
                                             questionnaire_width) %>%
  unique() -> setup_overview)

## save this table and anonymize by hand
write.csv(setup_overview,here::here("data", "transient_data_files", "setup_overview_all.csv"), 
            row.names = FALSE)

```

Load manually cleaned overview and create table

```{r}
setups <- read.csv2(here::here("data", "transient_data_files", 
                              "setup_overview_cleaned.csv"))
setups %>% group_by(OS) %>% count()
setups %>%
  mutate(group=case_when(subject<200 ~ "PWA", subject >200 ~"Control")) %>%
  group_by(group,browser) %>% count()
setups %>% group_by(system_format) %>% count()
setups %>% group_by(test_location) %>% count()
setups %>% group_by(screen_width) %>% count()
setups %>% 
  mutate(screen_width=case_when(screen_width=="1366 (1280 on day 8)" ~ 1366,
                                TRUE ~ as.numeric(as.character(screen_width)))) %>%
  summarise(min=min(screen_width), max=max(screen_width),
            mean=mean(screen_width), sd=sd(screen_width))
setups %>% group_by(screen_height) %>% count()
setups %>% 
  mutate(screen_height=case_when(screen_height=="768 (720  on day8)" ~ 768,
                                TRUE ~ as.numeric(as.character(screen_height)))) %>%
  summarise(min=min(screen_height), max=max(screen_height),
            mean=mean(screen_height), sd=sd(screen_height))
setups %>% group_by(questionnaire_width) %>% count()
```

## Fully anonymize data and reduce data frame

```{r anonymize}
# df_a <- df %>% dplyr::select(!"gender" & !starts_with("language") &
#                                !starts_with("handedness") &
#                                !starts_with("fingers") &
#                                 !starts_with("KB") &
#                                !starts_with("CH0") &
#                                !"comments" & !"type" &
#                                #!"gender_char" &
#                                !"itemvsnonitem1" &
#                                !"itemvsnonitem2" &
#                                !starts_with("MC") &
#                                !starts_with("SD") &
#                                !c("os_system", "os_system_raw",
#                                   "browser", "screen_width",
#                                   "system_format",
#                                   "questionnaire_width",
#                                   "screen_height", "OR02_01", "array") & 
#                                ! "age"
#                              # !"name" & !"time_wo_outlier"
#                                )
```


```{r write_anonymized_df}
#write.csv(df_a, here::here("data","data_long_anonymous.csv"))
```
